# Fix for issue the-darkvoid#12 (performance regression)