… BrcmPatchRAM/Info.plist
… README.md