It seems to be working well for me here.  The final addition of the IOSleep calls seem to fix some intermittent problems I was having (where firmware was loaded but native BT driver didn't work).  The IOSleep calls are conservative (10ms each) compared to the original code (5ms, 100ms, 50ms).

Please test and provide feedback.

Link to performance issue: https://github.com/the-darkvoid/BrcmPatchRAM/issues/12