@RehabMan 2015-03-14
I had another occurrence of bluetooth not starting after wake from sleep.  Not sure if it is a new problem or old problem.  I'll let you know when I figure it out...