@RehabMan 2015-03-16
Tracking the issue of intermittent lack of probe after sleep/wake cycle here: https://github.com/RehabMan/OS-X-BrcmPatchRAM/issues/2

It will take a few days to test the latest solution...