@RehabMan 2015-03-17
This appears to be solid now.  No case of bluetooth failure after wake from sleep in more than 48 hours of testing.  And bluetooth is available immediately after sleep without delay.